Update standalone-and-hvm.md
When I want to read the networking settings for my new hvm the panel in "qube manager" basic tab it shows the netmask "255.255.255.255". Here in this picture it shows "255.255.255.0". I don`t know whats correct and networking wont work for me. Is this picture wrong or the window in the os?